Add seed generator using hash mixing function ripped from murmurhash
Thanks to the great work by GitHub user @bryc, there’s a number of
different options for simple string to integer conversion using modified
hash functions. I’ve called the algorithm here `murmix`. It generates an
array of seed values by default—this provides for the possible future
addition of PRNG algorithms that require multiple seeds rather than a
single value as a starting point.

Added a basic first cut of the `seed` API, exported as a function at the
top level. I’m not sure if I will go ahead with this API as-is, as the
idea of having separate functions called `seed` and `seeds` to
distinguish between the array and a single value is quite appealing.

I also want to reconsider the way this package is structured. Tucking
everything inside the `src` directory is nice and consistent but it
doesn’t make for a great experience installing and importing from ES6
code. This also raises some questions around whether or not to still
care about CommonJS.